Added Live Match Data with Event Name
Added live match team name, as well as event name displayed. BUG TO FIX: When trying to retrieve the game and map score for each live match, beautifulsoup is not recognizing the string variables. Sometimes it will display the strings, but other times it will not. May need to switch my approach on how I am nesting the loops.